Resolve compiler warnings

+++ ASM/c/obj_kibako.c
@@ -4,7 +4,7 @@
 #include "textures.h"
 
 #define SMALLCRATE_DLIST (z64_gfx_t *)0x05005290
-#define SMALLCRATE_TEXTURE (uint64_t *)0x05011CA0
+#define SMALLCRATE_TEXTURE (uint8_t *)0x05011CA0
 extern bool POTCRATE_TEXTURES_MATCH_CONTENTS;
 
 override_t get_smallcrate_override(z64_actor_t *actor, z64_game_t *game)
